// app/types.ts
export interface TeamMember {
    name: string;
    role: string;
    image: string;
  }
  
  export interface Product {
    title: string;
    description: string;
    icon: string;
  }
  
  export interface Project {
    title: string;
    category: string;
    image: string;
    description: string;
    tech: string[];
  }
  